backend/requirements.txt
fastapi[all]
uvicorn 
pydantic
python-dotenv
openai